requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
# http://bethgelab.org/media/uploads/pytorch_models/vgg_conv.pth
aiogram==2.9.2
aiohttp==3.6.2
aiohttp-socks==0.3.9
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
certifi==2020.4.5.2
chardet==3.0.4
idna==2.9
multidict==4.7.6
numpy==1.18.5
Pillow==7.1.2
pytz==2020.1
torch==1.5.1+cpu
torchvision==0.6.1+cpu
yarl==1.4.2